implementacion de funcionalidad para el carrito de compras

+++ src/main/java/com/proyecto/ecommerce/model/DetalleOrden.java
@@ -18,7 +18,7 @@ public class DetalleOrden {
     private Integer id;
     
     private String nombre;
-    private Double cantidad;
+    private Integer cantidad;
     private Double precio;
     private Double total;
     
@@ -32,7 +32,7 @@ public class DetalleOrden {
         
     }
     
-    public DetalleOrden(Integer id, String nombre, Double cantidad, Double precio, Double total) {
+    public DetalleOrden(Integer id, String nombre, Integer cantidad, Double precio, Double total) {
         this.id = id;
         this.nombre = nombre;
         this.cantidad = cantidad;
@@ -56,11 +56,11 @@ public class DetalleOrden {
         this.nombre = nombre;
     }
 
-    public Double getCantidad() {
+    public Integer getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(Double cantidad) {
+    public void setCantidad(Integer cantidad) {
         this.cantidad = cantidad;
     }
